[refactor]: adicionar nova pasta em components para componentes especificos da página de protocolo + adicionar, remover e atualizar classes do componente LocationInput
