+++ src/components/protocol/DateInput.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import iconTime from '../assets/images/iconTime.svg';
+import iconDate from '../../assets/images/iconDate.svg';
 
 const styles = `
     .font-barlow {
@@ -26,39 +26,40 @@ const styles = `
         font-size: 1.1rem !important;
     }
 
-    .time-icon{
+    .date-icon {
         max-width: 50px;
     }
 `;
 
-function TimeInput(props) {
-    const currentTime = () => {
+function DateInput(props) {
+    const currentDate = () => {
         const date = new Date();
-        const hour = String(date.getHours()).padStart(2, '0');
-        const minutes = String(date.getMinutes()).padStart(2, '0');
-        return hour + ':' + minutes;
+        const day = String(date.getDate()).padStart(2, '0');
+        const month = String(date.getMonth() + 1).padStart(2, '0');
+        const year = String(date.getFullYear());
+        return year + '-' + month + '-' + day;
     };
 
     return (
         <div className="rounded-4 shadow bg-white overflow-hidden font-barlow p-0">
             <div className="row overflow-hidden m-0">
                 <div className="col-2 d-flex bg-pastel-blue p-0">
-                    <div className="time-icon ratio ratio-1x1 align-self-center w-50 mx-auto">
-                        <img src={iconTime} alt="Ícone de relógio" />
+                    <div className="date-icon ratio ratio-1x1 align-self-center w-50 mx-auto">
+                        <img src={iconDate} alt="Ícone de calendário" />
                     </div>
                 </div>
                 <div className="col p-3">
                     <div className="row m-0 pb-1">
-                        <label labelfor="timeinput" className="form-label color-dark-gray font-century-gothic fw-bold fs-7 m-0 p-0">
-                            Horário da coleta
+                        <label htmlFor="dateinput" className="form-label color-dark-gray font-century-gothic fw-bold fs-7 m-0 p-0">
+                            Data da coleta
                         </label>
                     </div>
                     <div className="row m-0">
                         <input
-                            type="time"
+                            type="date"
                             className="form-control border-0 color-sonic-silver fw-medium fs-7 w-auto m-0 p-0"
-                            id="timeinput"
-                            defaultValue={currentTime()}
+                            id="dateinput"
+                            defaultValue={currentDate()}
                         ></input>
                     </div>
                 </div>
@@ -68,4 +69,4 @@ function TimeInput(props) {
     );
 }
 
-export default TimeInput;
+export default DateInput;
