+++ src/components/protocol/TimeInput.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import iconDate from '../assets/images/iconDate.svg';
+import iconTime from '../../assets/images/iconTime.svg';
 
 const styles = `
     .font-barlow {
@@ -26,40 +26,39 @@ const styles = `
         font-size: 1.1rem !important;
     }
 
-    .date-icon {
+    .time-icon{
         max-width: 50px;
     }
 `;
 
-function DateInput(props) {
-    const currentDate = () => {
+function TimeInput(props) {
+    const currentTime = () => {
         const date = new Date();
-        const day = String(date.getDate()).padStart(2, '0');
-        const month = String(date.getMonth() + 1).padStart(2, '0');
-        const year = String(date.getFullYear());
-        return year + '-' + month + '-' + day;
+        const hour = String(date.getHours()).padStart(2, '0');
+        const minutes = String(date.getMinutes()).padStart(2, '0');
+        return hour + ':' + minutes;
     };
 
     return (
         <div className="rounded-4 shadow bg-white overflow-hidden font-barlow p-0">
             <div className="row overflow-hidden m-0">
                 <div className="col-2 d-flex bg-pastel-blue p-0">
-                    <div className="date-icon ratio ratio-1x1 align-self-center w-50 mx-auto">
-                        <img src={iconDate} alt="Ícone de calendário" />
+                    <div className="time-icon ratio ratio-1x1 align-self-center w-50 mx-auto">
+                        <img src={iconTime} alt="Ícone de relógio" />
                     </div>
                 </div>
                 <div className="col p-3">
                     <div className="row m-0 pb-1">
-                        <label htmlFor="dateinput" className="form-label color-dark-gray font-century-gothic fw-bold fs-7 m-0 p-0">
-                            Data da coleta
+                        <label labelfor="timeinput" className="form-label color-dark-gray font-century-gothic fw-bold fs-7 m-0 p-0">
+                            Horário da coleta
                         </label>
                     </div>
                     <div className="row m-0">
                         <input
-                            type="date"
+                            type="time"
                             className="form-control border-0 color-sonic-silver fw-medium fs-7 w-auto m-0 p-0"
-                            id="dateinput"
-                            defaultValue={currentDate()}
+                            id="timeinput"
+                            defaultValue={currentTime()}
                         ></input>
                     </div>
                 </div>
@@ -69,4 +68,4 @@ function DateInput(props) {
     );
 }
 
-export default DateInput;
+export default TimeInput;
